# frontend-materio/src/components/api/process-openapi.yaml
---
paths:
  '/processes/{id}/reviewprocess':
    summary: ReviewProcess operation on processes
    put:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                processId:
                  description: processId of this ReviewProcess
                  type: integer
                reviewResult:
                  description: reviewResult of this ReviewProcess
                  type: string
            examples: null
      parameters:
        - name: id
          description: ReviewProcess name
          schema:
            type: integer
          in: path
          required: true
          examples: null
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
              examples:
                '1':
                  value:
                    processId: 1
                    processName: A string
                    bpmnDefinition: A string
                    messages: value
          description: ReviewProcess Operation
      operationId: ReviewProcess
      summary: ReviewProcess
      description: ReviewProcess
  '/processes/{id}/modifyprocess':
    summary: ModifyProcess operation on processes
    put:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                processId:
                  description: processId of this ModifyProcess
                  type: integer
                newProcessName:
                  description: newProcessName of this ModifyProcess
                  type: string
                newTaskList:
                  description: newTaskList of this ModifyProcess
                  type: string
                newPermissions:
                  description: newPermissions of this ModifyProcess
                  type: string
            examples: null
      parameters:
        - name: id
          description: ModifyProcess name
          schema:
            type: integer
          in: path
          required: true
          examples: null
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
              examples:
                '1':
                  value:
                    processId: 1
                    processName: A string
                    bpmnDefinition: A string
                    messages: value
          description: ModifyProcess Operation
      operationId: ModifyProcess
      summary: ModifyProcess
      description: ModifyProcess
  '/processes/{id}/':
    summary: CreateProcess operation on processes
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                processName:
                  description: processName of this CreateProcess
                  type: string
                taskList:
                  description: taskList of this CreateProcess
                  type: string
                permissions:
                  description: permissions of this CreateProcess
                  type: string
            examples: null
      parameters:
        - name: id
          description: CreateProcess name
          schema:
            type: integer
          in: path
          required: true
          examples: null
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
              examples:
                '1':
                  value:
                    processId: 1
                    processName: A string
                    bpmnDefinition: A string
                    messages: value
          description: CreateProcess Operation
      operationId: CreateProcess
      summary: CreateProcess
      description: CreateProcess
  /processes:
    summary: Global operations on processes
    get:
      tags:
        - processes
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Process'
              examples:
                processes_json:
                  value:
                    - processId: 1
                      processName: A string
                      bpmnDefinition: A string
                      messages: value
          description: Get list of processes
      operationId: GetProcess
      summary: Get list of processes
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Process'
            examples:
              '1':
                value:
                  processId: 1
                  processName: A string
                  bpmnDefinition: A string
                  messages: value
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
              examples:
                '1':
                  value:
                    processId: 1
                    processName: A string
                    bpmnDefinition: A string
                    messages: value
          description: Changed Process
      operationId: Post Process
      summary: Post Process
  '/processes/{id}':
    summary: Specific operation on processes
    get:
      parameters:
        - examples:
            '1':
              value: 1
          name: id
          description: Process name
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
              examples:
                '1':
                  value:
                    processId: 1
                    processName: A string
                    bpmnDefinition: A string
                    messages: value
          description: processes with specified id
      operationId: GetProcessById
      summary: Get Process by id
      description: Get Process by id
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Process'
            examples:
              '1':
                value:
                  processId: 1
                  processName: A string
                  bpmnDefinition: A string
                  messages: value
        required: true
      parameters:
        - examples:
            '1':
              value: 1
          name: processId
          description: Process name
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Process'
              examples:
                '1':
                  value:
                    processId: 1
                    processName: A string
                    bpmnDefinition: A string
                    messages: value
          description: Changed Process
      operationId: PatchProcess
      summary: Patch existing Process
    parameters:
      - name: id
        description: Process id
        schema:
          type: integer
        in: path
        required: true
openapi: 3.0.2
info:
  title: Process
  version: 0.0.1
  description: API definition of process
  contact:
    name: creator's name
    url: 'http://creator.homepage.address'
    email: creator@creator.com
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
components:
  schemas:
    Process:
      title: Root Type for Process
      description: The root of the Process type's schema.
      type: object
      properties:
        processId:
          description: processId of this Process
          type: integer
        processName:
          description: processName of this Process
          type: string
        bpmnDefinition:
          description: bpmnDefinition of this Process
          type: string
        messages:
          description: messages of this Process
          type: string
tags:
  - name: Process
    description: Process resource